Desktop/Work/Airbnb/ml_deployment/flask_apps/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
numpy==1.13.3
keras==2.2.4
Pillow==5.2.0
tensorflow==1.13.1
